# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Money Manager Pro API
  version: 1.0.0-draft
servers:
  - url: /api/v1
paths:
  /transactions:
    get:
      operationId: listTransactions
      summary: List transactions
      description: Returns the caller's transactions in integer minor units. Existing records must first receive stable IDs through `mm-admin migrate-transaction-ids`.
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: account
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: sort
          in: query
          schema:
            type: string
            enum: [date, amount]
            default: date
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Transactions owned by the caller
          content:
            application/json:
              schema:
                type: object
                required: [transactions]
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiTransaction'
              example:
                transactions:
                  - id: tx_01234567-89ab-cdef-0123-456789abcdef
                    account: Daily
                    amountMinor: -1250
                    currency: EUR
                    date: '2026-09-06'
                    time: '09:30'
                    category: '@Groceries'
                    comment: Weekly shop
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/InsufficientScope'
    post:
      operationId: createTransaction
      summary: Create a transaction
      description: Adds a transaction and atomically rebuilds its derived income, expense, Smile, Fire, and Mojo state.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransactionRequest'
            example:
              account: Daily
              amountMinor: -1250
              date: '2026-09-06'
              time: '09:30'
              category: '@Groceries'
              comment: Weekly shop
      responses:
        '201':
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiTransaction'
              example:
                id: tx_01234567-89ab-cdef-0123-456789abcdef
                account: Daily
                amountMinor: -1250
                currency: EUR
                date: '2026-09-06'
                time: '09:30'
                category: '@Groceries'
                comment: Weekly shop
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/InsufficientScope'
  /transactions/export:
    get:
      operationId: exportTransactions
      summary: Export all transactions as newline-delimited JSON
      description: Returns every transaction owned by the caller, one JSON object per line (no pagination, no filters), for full-account moves. Requires `transactions:bulk`.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Newline-delimited JSON, one ApiTransaction per line
          content:
            application/x-ndjson:
              schema:
                type: string
              example: |-
                {"id":"tx_01234567-89ab-cdef-0123-456789abcdef","account":"Daily","amountMinor":-1250,"currency":"EUR","date":"2026-09-06","time":"09:30","category":"@Groceries","comment":"Weekly shop"}
                {"id":"tx_11111111-89ab-cdef-0123-456789abcdef","account":"Income","amountMinor":250000,"currency":"EUR","date":"2026-09-01","time":"08:00","category":"@Salary","comment":""}
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/InsufficientScope'
  /transactions/import:
    post:
      operationId: importTransactions
      summary: Bulk-create transactions from newline-delimited JSON
      description: >-
        Creates one transaction per line of the request body (each line the same shape as CreateTransactionRequest).
        Requires `transactions:bulk` and an `Idempotency-Key` header — hashed from the raw request body, not a
        re-serialized JSON value like batch, so byte-identical replay is required (see docs/api/AGENTS.md) — with
        the same replay/mismatch contract and the same known write-then-record ordering gap as
        `POST /transactions/batch`. Up to 10,000 lines of at most 16KB each per call. `?atomic=true` makes the whole
        import all-or-nothing, same as batch's `atomic` flag; the default applies what parses and validates,
        reporting the rest per line as `error`.
      security:
        - bearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
        - name: atomic
          in: query
          schema:
            type: boolean
            default: false
      requestBody:
        required: true
        content:
          application/x-ndjson:
            schema:
              type: string
            example: |-
              {"account":"Daily","amountMinor":-850,"date":"2026-09-07","time":"08:15","category":"@Coffee","comment":""}
              {"account":"Income","amountMinor":250000,"date":"2026-09-01","time":"08:00","category":"@Salary","comment":""}
      responses:
        '200':
          description: Per-line results, in the same shape as a batch response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchTransactionsResponse'
              example:
                idempotencyKey: 2026-09-07-restore
                atomic: false
                itemCount: 2
                results:
                  - op: create
                    status: created
                    id: tx_22222222-89ab-cdef-0123-456789abcdef
                    transaction:
                      id: tx_22222222-89ab-cdef-0123-456789abcdef
                      account: Daily
                      amountMinor: -850
                      currency: EUR
                      date: '2026-09-07'
                      time: '08:15'
                      category: '@Coffee'
                      comment: ''
                  - op: create
                    status: created
                    id: tx_33333333-89ab-cdef-0123-456789abcdef
                    transaction:
                      id: tx_33333333-89ab-cdef-0123-456789abcdef
                      account: Income
                      amountMinor: 250000
                      currency: EUR
                      date: '2026-09-01'
                      time: '08:00'
                      category: '@Salary'
                      comment: ''
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/InsufficientScope'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
  /transactions/batch:
    post:
      operationId: batchTransactions
      summary: Create, update, or delete many transactions in one request
      description: >-
        Applies a mix of create/update/delete operations atomically at the storage level (one document write) and
        rebuilds derived income/expense/Smile/Fire/Mojo state once. Requires the `transactions:bulk` scope and an
        `Idempotency-Key` header — replaying the same key with the same body returns the original result without
        reapplying it; replaying it with a different body is rejected. The write and the idempotency record are not
        one atomic transaction: a crash between them (rare) can make a same-key retry reapply the batch — see
        docs/api/AGENTS.md. Setting `atomic` to true makes the whole batch all-or-nothing (if any operation fails,
        none are applied, and a would-have-succeeded `create`'s `id` in the response was never persisted); the
        default, false, applies what succeeds and reports failures per item.
      security:
        - bearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchTransactionsRequest'
            example:
              atomic: false
              operations:
                - op: create
                  account: Daily
                  amountMinor: -850
                  date: '2026-09-07'
                  time: '08:15'
                  category: '@Coffee'
                  comment: ''
                - op: update
                  id: tx_01234567-89ab-cdef-0123-456789abcdef
                  amountMinor: -1500
                - op: delete
                  id: tx_11111111-89ab-cdef-0123-456789abcdef
      responses:
        '200':
          description: Per-item results (some items may report `error` even on a 200, in non-atomic mode)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchTransactionsResponse'
              example:
                idempotencyKey: 2026-09-07-morning-batch
                atomic: false
                itemCount: 3
                results:
                  - op: create
                    status: created
                    id: tx_22222222-89ab-cdef-0123-456789abcdef
                    transaction:
                      id: tx_22222222-89ab-cdef-0123-456789abcdef
                      account: Daily
                      amountMinor: -850
                      currency: EUR
                      date: '2026-09-07'
                      time: '08:15'
                      category: '@Coffee'
                      comment: ''
                  - op: update
                    status: error
                    id: tx_01234567-89ab-cdef-0123-456789abcdef
                    error: No matching transaction exists.
                  - op: delete
                    status: deleted
                    id: tx_11111111-89ab-cdef-0123-456789abcdef
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/InsufficientScope'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
  /transactions/{transactionId}/copy:
    post:
      operationId: copyTransaction
      summary: Copy a transaction
      description: Creates a new transaction from an existing one's account, amount, category, and comment. Date and time default to now, matching the UI's copy behavior; any field may be overridden in the request body. Atomically rebuilds derived income, expense, Smile, Fire, and Mojo state.
      security:
        - bearerAuth: []
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CopyTransactionRequest'
            example:
              date: '2026-09-07'
      responses:
        '201':
          description: The newly created copy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiTransaction'
              example:
                id: tx_11111111-89ab-cdef-0123-456789abcdef
                account: Daily
                amountMinor: -1250
                currency: EUR
                date: '2026-09-07'
                time: '09:30'
                category: '@Groceries'
                comment: Weekly shop
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/InsufficientScope'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ApiTransaction:
      type: object
      required: [id, account, amountMinor, currency, date, time, category, comment]
      properties:
        id:
          type: string
        account:
          type: string
        amountMinor:
          type: integer
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
        date:
          type: string
          format: date
        time:
          type: string
        category:
          type: string
        comment:
          type: string
    CreateTransactionRequest:
      type: object
      required: [account, amountMinor, date, time, category, comment]
      properties:
        account:
          type: string
        amountMinor:
          type: integer
          not:
            enum: [0]
          description: Integer minor units. Must not be zero.
        date:
          type: string
          format: date
        time:
          type: string
        category:
          type: string
        comment:
          type: string
...
    Problem:
      type: object
      required: [type, title, status, detail, code]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        code:
          type: string
  responses:
    AuthenticationRequired:
      description: Authentication required
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://money-manager.dev/problems/auth_required
            title: Authentication required
            status: 401
            detail: Provide a session or PAT.
            code: auth_required
...
    InsufficientScope:
      description: The supplied token does not grant the required scope
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://money-manager.dev/problems/scope_insufficient
            title: Insufficient scope
            status: 403
            detail: This operation requires the 'transactions:r' scope.
            code: scope_insufficient
...